backend/requirements.txt
# ───── Web framework ─────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# ───── Settings / validation ─────
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0

# ───── DB ─────
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# ───── Auth ─────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ───── MQTT ─────
asyncio-mqtt==0.16.2
paho-mqtt==1.6.1

# ───── Modbus ─────
pymodbus==3.7.4

# ───── Scheduler ─────
apscheduler==3.11.0

# ───── HTTP client (notifications) ─────
httpx==0.28.1

# ───── Logging ─────
structlog==24.4.0

# ───── Redis ─────
redis==5.2.1

# ───── Excel parsing (manual upload) ─────
openpyxl==3.1.5

# ───── Testing ─────
pytest==8.3.4
pytest-asyncio==0.25.2
